Add NocoDB user account

Adds a NocoDB user account on a NocoDB server without any user interaction. E-mail invitation is bypassed and the user's e-mail address is automatically validated. A convenience function that combines several of the other functions found in this package.

Usage

add_user(
  user_email,
  user_password,
  display_name = NULL,
  subscribe_to_newsletter = FALSE,
  origin = funky::config_val("origin"),
  email = funky::config_val("email", require = TRUE),
  password = funky::config_val("password", require = TRUE),
  quiet = TRUE
)

Arguments

user_email

character(1)
E-mail address of the user to sign up.

user_password

character(1)
Password of the user to sign up.

display_name

character(1)
Name to be displayed for the user in NocoDB. Omitted if NULL.

subscribe_to_newsletter

logical(1)
Whether or not to subscribe the signed up user to the NocoDB newsletter.

origin

character(1)
NocoDB server origin.

email

character(1)
E-mail address of the NocoDB user to authenticate with.

password

character(1)
Password of the NocoDB user to authenticate with.

quiet

logical(1)
Whether or not to suppress printing status output from internal processing.

Value

A tibble with metadata about the updated NocoDB user, invisibly.

Details

add_user() does not support authentication via API tokens because of update_app_settings().

See also