Get NocoDB data source ID

Returns the identifier of the data source with the specified alias in the specified base on a NocoDB server.

Usage

data_src_id(
  alias,
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)

Arguments

alias

character(1)
Alias of the data source.

id_base

character(1)
NocoDB base identifier as returned by base_id().

origin

character(1)
NocoDB server origin.

email

character(1)
E-mail address of the NocoDB user to authenticate with.

password

character(1)
Password of the NocoDB user to authenticate with.

api_token

character(1)
NocoDB API token. Takes precedence over email and password if provided.

Value

Data source identifier as a character scalar.

See also