Decode access token (JWT)

Decodes an access token that adheres to the JSON Web Token (JWT) standard and returns it as a tibble. Only the payload is returned.

Usage

decode_access_token(x)

Arguments

x

Access token to be decoded.

Value

A tibble.

Column iat gives the time at which the JWT was issued and column exp gives the expiration time on or after which the JWT must not be accepted anymore. They are registered claim names in the JWT standard.

See also