Get NocoDB integration ID

Returns the identifier of the integration with the specified type, sub_type and/or title on a NocoDB server.

Usage

integration_id(
  title = NULL,
  type = "database",
  sub_type = c("mysql2", "pg", "sqlite3"),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)

Arguments

title

pkgsnip::type("chr") NocoDB integration title. NULL means any title.

type

pkgsnip::type("chr") NocoDB integration type. One of "database", or NULL for any type.

sub_type

pkgsnip::type("chr") NocoDB integration subtype. One of "mysql2", "pg" or "sqlite3", or NULL for any subtype.

hostname

NocoDB server hostname. A character scalar.

email

E-mail address of the NocoDB user to authenticate with.

password

Password of the NocoDB user to authenticate with.

api_token

NocoDB API token. Takes precedence over email and password if provided.

Value

A character scalar.

Details

Note that NocoDB allows to create multiple integrations with the same title, type and sub_type, so integration_id() cannot discern such cases and just returns the identifier of the first integration listed in the API response (with a suitable warning).

See also