List NocoDB integrations

Returns a tibble with metadata about all configured integrations on a NocoDB server from its GET /api/v2/meta/integrations API endpoint.

Usage

integrations(
  type = NULL,
  id_base = NULL,
  incl_config = TRUE,
  decode_config = TRUE,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)

Arguments

type

character(1)
Optional integration type to limit the results to. One of "database".

id_base

character(1)
Optional NocoDB base identifier (as returned by base_id()) to limit the results to.

incl_config

logical(1)
Whether or not to include the integration configuration in the result.

decode_config

logical(1)
Whether or not to decode the integration configuration in the result. If FALSE, the whole configuration is returned as a single Base64-encoded and encrypted config column, otherwise it is returned as multiple clear-text config.* columns. Only relevant if incl_config = TRUE.

origin

NocoDB server origin. A character scalar.

email

E-mail address of the NocoDB user to authenticate with.

password

Password of the NocoDB user to authenticate with.

api_token

NocoDB API token. Takes precedence over email and password if provided.

Value

A tibble.

See also