Re-order NocoDB table

Sets the numeric order of the specified table on a NocoDB server via its POST /api/v2/meta/tables/{tbl_id}/reorder API endpoint.

Usage

reorder_tbl(
  tbl_id = tbl_id(),
  order = 1L,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)

Arguments

tbl_id

NocoDB table identifier as returned by tbl_id(). A character scalar.

order

A number to assign as the table's order "weight".

hostname

NocoDB server hostname. A character scalar.

auth_token

NocoDB API authentication token. A character scalar.

Value

tbl_id, invisibly.

Details

Lower numbers place the table higher up in the UI and vice versa. The current order of all the tables in a base can be determined via tbls().

See also