Set metadata for NocoDB tables

Sets the provided table metadata on a NocoDB server. Currently, this includes:

Setting the order of the tables in the base using reorder_tbl() according to the row order of data.

Setting the table icons (emojis) using update_tbl() according to data$meta.icon.

Usage

set_tbl_metadata(
  data,
  base_id = base_id(),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  quiet = FALSE
)

Arguments

data

Dataframe with the columns name, meta.icon that defines the table-name-and-metadata-value mapping to be applied. Additional columns are ignored.

base_id

NocoDB base identifier as returned by base_id(). A character scalar.

hostname

NocoDB server hostname. A character scalar.

auth_token

NocoDB API authentication token. A character scalar.

quiet

Whether or not to suppress printing status output from internal processing.

Value

NULL, invisibly.

See also