Sign out NocoDB user

Invalidates a cached refresh token via the POST /api/v1/auth/user/signout API endpoint and clears the associated cache on disk.

Usage

sign_out(
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email", require = TRUE),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token"),
  quiet = FALSE
)

Arguments

origin

NocoDB server origin. A character scalar.

email

E-mail address of the NocoDB user to sign out. A character scalar.

password

Password of the NocoDB user to authenticate with.

api_token

NocoDB API token. Takes precedence over email and password if provided.

quiet

Whether or not to suppress printing status output from internal processing.

Value

email, invisibly.

Details

Note that the email address to sign out and the api_token to authenticate do not necessarily need to belong to the same user.

See also