Sign up NocoDB user

Signs up an invited user on a NocoDB server via its POST /api/v1/auth/user/signup API endpoint.

Usage

sign_up_user(
  email_new,
  password_new,
  invite_token = NULL,
  subscribe_to_newsletter = TRUE,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, required = FALSE),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg, required = FALSE),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg, required = FALSE)
)

Arguments

email_new

E-mail address of the user to sign up. A character scalar.

password_new

Password of the user to sign up. A character scalar.

invite_token

Invite token of the user to sign up. A character scalar.

subscribe_to_newsletter

Whether or not to subscribe the signed up user to the NocoDB newsletter.

hostname

NocoDB server hostname. A character scalar.

email

E-mail address of the NocoDB user to authenticate with.

password

Password of the NocoDB user to authenticate with.

api_token

NocoDB API token. Takes precedence over email and password if provided.

Value

The generated access token for the signed up user as a character scalar, invisibly.

API errors

If you did provide an invite_token and encounter a Not allowed to signup error, it means the provided email_new does not match the invite_token.

If you did not provide an invite_token and encounter a Not allowed to signup error, it means sign-up of new users is restricted to invitees only.

If you encounter an Invalid invite url error, it means the provided invite_token is either invalid or has already been used to sign up.

See also