Synchronize NocoDB data source schema eagerly

Synchronizes the specified data source's schema between its external and its NocoDB-internal state by repeatedly calling sync_data_src() until has_data_src_diff() doesn't detect any more changes.

Usage

sync_data_src_eagerly(
  id_data_src,
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val("hostname"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token"),
  wait_max = 30L,
  wait_resync = 7
)

Arguments

id_data_src

NocoDB data source identifier as returned by data_src_id(). A character scalar.

id_base

NocoDB base identifier as returned by base_id(). A character scalar.

hostname

NocoDB server hostname. A character scalar.

email

E-mail address of the NocoDB user to authenticate with.

password

Password of the NocoDB user to authenticate with.

api_token

NocoDB API token. Takes precedence over email and password if provided.

wait_max

Maximum time in seconds to wait for schema synchronization to finish. An error is thrown if there are still schema changes pending when the timeout is reached. An integerish scalar >= wait_resync.

wait_resync

Interval in seconds to repeat calling sync_data_src(). A number >= 1 and <= wait_max.

Value

id_data_src, invisibly.