Update NocoDB base visibility rules

Updates the visibility rules (UI ACL) for the different views of the specified table view on a NocoDB server via its POST /api/v2/meta/bases/{id_base}/visibility-rules API endpoint.

Usage

update_base_ui_acl(
  id_tbl_view,
  guest = viewer,
  viewer = commenter,
  commenter = editor,
  editor = creator,
  creator = owner,
  owner = FALSE,
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val("hostname"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token"),
  quiet = FALSE
)

Arguments

id_tbl_view

Table view identifier as returned by tbl_view_id().

guest

Whether or not to hide the table view for users with role guest.

viewer

Whether or not to hide the table view for users with role viewer.

commenter

Whether or not to hide the table view for users with role commenter.

editor

Whether or not to hide the table view for users with role editor.

creator

Whether or not to hide the table view for users with role creator.

owner

Whether or not to hide the table view for users with role owner.

id_base

NocoDB base identifier as returned by base_id(). A character scalar.

hostname

NocoDB server hostname. A character scalar.

email

E-mail address of the NocoDB user to authenticate with.

password

Password of the NocoDB user to authenticate with.

api_token

NocoDB API token. Takes precedence over email and password if provided.

quiet

Whether or not to suppress printing status output from internal processing.

Value

id_tbl_view, invisibly.

See also