Update NocoDB data source

Updates the specified data source of the specified base on a NocoDB server via its PATCH /api/v2/meta/bases/{id_base}/sources/{id_data_src} API endpoint.

Usage

update_data_src(
  id_data_src,
  connection = NULL,
  type = NULL,
  alias = NULL,
  inflection_column = NULL,
  inflection_table = NULL,
  is_schema_readonly = NULL,
  is_data_readonly = NULL,
  enabled = NULL,
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)

Arguments

id_data_src

character(1)
NocoDB data source identifier as returned by data_src_id().

connection

list()
Type-specific connection details for the data source.

type

character(1)
Type of the data source. One of "mssql", "mysql", "pg" or "sqlite3".

alias

character(1)
Alias of the data source.

inflection_column

character(1)
Type of inflection to apply for column names in the data source. One of "none" or "camelize".

inflection_table

character(1)
Type of inflection to apply for table names in the data source. One of "none" or "camelize".

is_schema_readonly

logical(1)
Whether or not to forbid alterations of the data source's schema from NocoDB.

is_data_readonly

logical(1)
Whether or not to forbid alterations of the data source's data from NocoDB.

enabled

logical(1)
Whether the added data source is enabled or disabled.

id_base

character(1)
NocoDB base identifier as returned by base_id().

origin

character(1)
NocoDB server origin.

email

character(1)
E-mail address of the NocoDB user to authenticate with.

password

character(1)
Password of the NocoDB user to authenticate with.

api_token

character(1)
NocoDB API token. Takes precedence over email and password if provided.

Value

A tibble with metadata about the updated NocoDB data source, invisibly.

See also