Update NocoDB table

Updates the metadata of the specified table on a NocoDB server via its PATCH /api/v2/meta/tables/{id_tbl} API endpoint.

Usage

update_tbl(
  id_tbl,
  id_base = NULL,
  name = NULL,
  title = NULL,
  description = NULL,
  meta = NULL,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token"),
  quiet = FALSE
)

Arguments

id_tbl

character(1)
NocoDB table identifier as returned by tbl_id().

id_base

character(1)
NocoDB base identifier as returned by base_id().

name

character(1)
Table name.

title

character(1)
NocoDB-specific table title.

description

character(1)
NocoDB table description displayed as a tooltip in the user interface. Omitted if NULL.

meta

list()
NocoDB-specific table metadata as a list or something coercible to like a data frame.

origin

character(1)
NocoDB server origin.

email

character(1)
E-mail address of the NocoDB user to authenticate with.

password

character(1)
Password of the NocoDB user to authenticate with.

api_token

character(1)
NocoDB API token. Takes precedence over email and password if provided.

quiet

logical(1)
Whether or not to suppress printing status output from internal processing.

Value

id_tbl, invisibly.

See also