Set metadata for NocoDB tables

Sets the provided table metadata on a NocoDB server. Currently, this includes:

Setting the order of the tables in the base using reorder_tbl() according to the row order of data.

Setting the table icons (emojis) using update_tbl() according to data$meta.icon.

Usage

set_tbl_metadata(
  data,
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val("hostname"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token"),
  quiet = FALSE
)

Arguments

data

Data frame with the columns name, meta.icon that defines the table-name-and-metadata-value mapping to be applied. Additional columns are ignored.

id_base

NocoDB base identifier as returned by base_id(). A character scalar.

hostname

NocoDB server hostname. A character scalar.

email

E-mail address of the NocoDB user to authenticate with.

password

Password of the NocoDB user to authenticate with.

api_token

NocoDB API token. Takes precedence over email and password if provided.

quiet

Whether or not to suppress printing status output from internal processing.

Value

NULL, invisibly.

See also